+++ docs/changelog.rst
@@ -5,6 +5,7 @@ Version 0.1.0
 --------------
 
 + Add support for pages
++ Build to tempdir when running ``catsup server``
 + Add ``config.site.description``
 + Use ``config.comment.shortname`` to replace ``config.comment.disqus`` and ``config.comment.duoshuo``
 + Regenerate the site when your theme or posts changed when running ``catsup server``